// api/src/models/NftMetadata.ts
export interface NftMetadata {
  standard: string;
  type: string;
  version: string;
  uri?: string;
  name: string;
  collectionName?: string;
  royalties?: { [key: string]: number };
  issuerName?: string;
  description?: string;
  attributes: [{
    eventIssuerAddress: string;
    eventId: string;
    ticketPrice: number;
    eventDate: string;
    isCollectionNft: boolean;
  }];
}